aim_backend/requirements.txt
aim-metrics==1.0.1
aim-segmentation==1.0
autopep8==1.3.3
backports-abc==0.5
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
flake8==3.5.0
futures==3.2.0
idna==2.7
importmagic==0.1.7
jedi==0.11.0
matplotlib==2.1.0
mccabe==0.6.1
motor==2.0.0
networkx==2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.4.2.17
pandas==0.23.4
parso==0.1.0
Pillow==5.3.0
pycodestyle==2.3.1
pyflakes==1.6.0
pymongo==3.7.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-resize-image==1.1.18
pytz==2017.3
PyWavelets==0.5.2
requests==2.19.1
scikit-image==0.13.1
scipy==1.0.0
selenium==3.7.0
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
tornado==4.5.2
urllib3==1.23
yapf==0.20.0
